Split the redirection tests out of dispatch/routing_test.rb into dispatch/routing/redirection_test.rb
Nest dispatch/routing/concerns_test consistent with the other routing test files.

Removed some unnecessary requires from dispatch/routing_test
